chore: rename to index.tsx
