+++ packages/frontend/src/components/VisualizationConfigs/PieChartConfig/ConfigTabs/index.tsx
